// Frontend/src/enums/AC/ac01.ts
export enum AC01Status {
  All = 'All',
  Draft = 'Draft',
  Rejected = "Rejected",
  Approved = 'Approved',
  Edit = 'Edit',
  WaitingApproval = 'WaitingApproval',
  WaitingForCompletion = 'WaitingForCompletion',
}

export enum sourceType {
  W119 = 'W119',
  Clause79_2 = 'Clause79_2',
  Disbursement = 'Disbursement',
  ContractGuaranteeReturn = 'ContractGuaranteeReturn',
  PettyCashReimbursement = 'PettyCashReimbursement'
}